move listener cache into instance doc
