+++ packages/xforms-engine/src/instance/PrimaryInstance.ts
@@ -48,6 +48,7 @@ import type { PrimaryInstanceDocument } from './internal-api/PrimaryInstanceDocu
 import type { ClientReactiveSerializableInstance } from './internal-api/serialization/ClientReactiveSerializableInstance.ts';
 import type { TranslationContext } from './internal-api/TranslationContext.ts';
 import { Root } from './Root.ts';
+import type { ValueChangedEventListener } from './internal-api/ValueChangedEventListener.ts';
 
 /**
  * As specified by {@link | XPath 1.0}:
@@ -129,6 +130,7 @@ export class PrimaryInstance<
   readonly initializationMode: FormInstanceInitializationMode;
   readonly model: ModelDefinition;
   readonly attachments: InstanceAttachmentsState;
+  readonly valueChangedEventListeners: Map<string, ValueChangedEventListener[]>;
 
   // InstanceNode
   protected readonly state: SharedNodeState<PrimaryInstanceStateSpec>;
@@ -185,9 +187,7 @@ export class PrimaryInstance<
 
     this.initializationMode = mode;
     this.model = model;
-    if (mode !== 'restore') {
-      this.model.restore(); // TODO does it make more sense for the listeners to be here?
-    }
+    this.valueChangedEventListeners = new Map();
 
     this.attachments = new InstanceAttachmentsState(initialState?.attachments, fetchFormAttachment);
     this.instanceNode = activeInstance;
@@ -254,6 +254,10 @@ export class PrimaryInstance<
     childrenState.setChildren([root]);
     this.attributeState.setAttributes(buildAttributes(this));
     setIsAttached(true);
+  }
+
+  getValueChangedEventListeners(): Map<string, ValueChangedEventListener[]> {
+    return this.valueChangedEventListeners;
   }
 
   override getAttributes(): readonly Attribute[] {
