+++ packages/xforms-engine/src/instance/internal-api/AttributeContext.ts
@@ -7,11 +7,13 @@ import type { AnyBindPreloadDefinition } from '../../parse/model/BindPreloadDefi
 import type { ModelDefinition } from '../../parse/model/ModelDefinition.ts';
 import type { EvaluationContext } from './EvaluationContext.ts';
 import type { InstanceConfig } from './InstanceConfig.ts';
+import type { ValueChangedEventListener } from './ValueChangedEventListener.ts';
 
 export interface InstanceAttributeContextDocument {
   readonly initializationMode: FormInstanceInitializationMode;
   readonly isAttached: Accessor<boolean>;
   getBackgroundGeopoint: Accessor<Promise<string>>;
+  getValueChangedEventListeners(): Map<string, ValueChangedEventListener[]>;
 }
 
 export type DecodeInstanceValue = (value: string) => string;
